data-model: move freeze-protocol plumbing off FabricInstance onto BaseFabricInstance
The `[DEEP_FREEZE]` / `[IS_DEEP_FROZEN]` symbol-keyed members are implementation
plumbing, not part of the pure `instanceof`-able protocol that external code is
written against. Move their `unique symbol` consts and abstract declarations off
`FabricInstance` (the dependency-free protocol leaf in interface.ts) down onto
`BaseFabricInstance`, where the concrete instance machinery already lives.
`deepClone()` / `shallowClone()` stay on `FabricInstance` as the client surface.

The generic freeze machinery in deep-freeze.ts gated on `instanceof
FabricInstance` before invoking those symbols; since the members no longer type
against `FabricInstance`, the three gates now narrow on `instanceof
BaseFabricInstance`. Every concrete instance already extends `BaseFabricInstance`,
so the gate is runtime-equivalent. Importers of the two symbols (subclass impls,
deep-freeze.ts, and tests) are re-pointed to the new home; all within
data-model, no external contract change.

Structural refactor with no runtime behavior change; the data-model suite passes
unchanged (test edits are import-path re-points only).
