+++ packages/data-model/src/deep-freeze.ts
@@ -1,13 +1,12 @@
 import { isPlainObject } from "@commonfabric/utils/types";
 import { isArrayWithOnlyIndexProperties } from "@commonfabric/utils/arrays";
 
+import { FabricPrimitive, FabricValue } from "./interface.ts";
 import {
+  BaseFabricInstance,
   DEEP_FREEZE,
-  FabricInstance,
-  FabricPrimitive,
-  FabricValue,
   IS_DEEP_FROZEN,
-} from "./interface.ts";
+} from "./fabric-instances/BaseFabricInstance.ts";
 
 /**
  * Cache of confirmed deep-frozen objects.
@@ -122,18 +121,18 @@ export function isDeepFrozen(value: unknown): boolean {
 
     let result = true;
 
-    if (obj instanceof FabricInstance) {
-      // A `FabricInstance`'s logical contents are not its enumerable own-props
+    if (obj instanceof BaseFabricInstance) {
+      // A fabric instance's logical contents are not its enumerable own-props
       // (e.g. a `FabricError` keeps its custom properties in a private extras
       // `Map`), so it answers the deep-frozen question via its
       // `[IS_DEEP_FROZEN]` protocol member -- the side-effect-free sibling of
       // `[DEEP_FREEZE]` -- recursing into each nested `FabricValue` through
       // `check`, which shares this call's cycle state. Gating on `instanceof`
-      // against the abstract base keeps this generic; the member is abstract on
-      // `FabricInstance`, so every instance implements it. (A `FabricPrimitive`
-      // is necessarily frozen with no outbound references, so the
-      // `Object.values` arm below answers it correctly by accident -- its empty
-      // enumerable props yield `true`.)
+      // against `BaseFabricInstance` keeps this generic; the member is abstract
+      // on `BaseFabricInstance`, so every instance implements it. (A
+      // `FabricPrimitive` is necessarily frozen with no outbound references, so
+      // the `Object.values` arm below answers it correctly by accident -- its
+      // empty enumerable props yield `true`.)
       result = obj[IS_DEEP_FROZEN](check);
     } else if (Array.isArray(obj)) {
       for (let i = 0; i < obj.length; i++) {
@@ -170,10 +169,10 @@ export function isDeepFrozen(value: unknown): boolean {
  *    objects): short-circuit unchanged.
  * 2. `FabricPrimitive` instance: short-circuit unchanged -- these self-freeze
  *    at construction and have no outbound references.
- * 3. `FabricInstance` (the abstract base): delegate generically to its
- *    `[DEEP_FREEZE]` protocol member, handing recursion through as the
- *    `subFreeze` callback. The dispatch gates on `instanceof` against the
- *    abstract base -- it operates generically and does not enumerate
+ * 3. Fabric instance: delegate generically to its `[DEEP_FREEZE]` protocol
+ *    member, handing recursion through as the `subFreeze` callback. The
+ *    dispatch gates on `instanceof` against `BaseFabricInstance` (where the
+ *    member is declared) -- it operates generically and does not enumerate
  *    concrete subclasses.
  * 4. Plain object or array: recursively freeze children, then freeze the
  *    container.
@@ -234,12 +233,12 @@ export function deepFreeze<T>(value: T): T {
     }
     inProgress.add(obj);
 
-    // Arm 3: a `FabricInstance` freezes itself in place via its `[DEEP_FREEZE]`
+    // Arm 3: a fabric instance freezes itself in place via its `[DEEP_FREEZE]`
     // protocol member. `freeze` is handed in as the `subFreeze` callback: it
     // closes over `inProgress`, so the impl's recursion into nested
     // `FabricValue`s shares cycle state with this call -- the participating
     // instance doesn't need to be `inProgress`-aware in its own signature.
-    if (value instanceof FabricInstance) {
+    if (value instanceof BaseFabricInstance) {
       const result = value[DEEP_FREEZE](freeze) as U;
       // Cache the now-deep-frozen result so subsequent `isDeepFrozen()` checks
       // short-circuit in O(1), mirroring arm 4's cache-write below.
@@ -331,16 +330,16 @@ export function isDeepFrozenFabricValue(value: unknown): value is FabricValue {
       // `FabricPrimitive`s are by definition frozen and have no outbound
       // references.
       return true;
-    } else if (item instanceof FabricInstance) {
-      // Object.freeze() cannot prove that a FabricInstance's private logical
+    } else if (item instanceof BaseFabricInstance) {
+      // Object.freeze() cannot prove that a fabric instance's private logical
       // contents will remain unchanged, so validate it but do not root-cache a
       // graph that reaches one.
       cacheableByIdentity = false;
-      // `FabricInstance`s answer the deep-frozen question via their
+      // Fabric instances answer the deep-frozen question via their
       // `[IS_DEEP_FROZEN]` protocol member (the side-effect-free sibling of
       // `[DEEP_FREEZE]`), recursing through `checkValue`. Gating on
-      // `instanceof` against the abstract base keeps this guard generic; the
-      // `[IS_DEEP_FROZEN]` member is abstract on `FabricInstance`, so every
+      // `instanceof` against `BaseFabricInstance` keeps this guard generic; the
+      // `[IS_DEEP_FROZEN]` member is abstract on `BaseFabricInstance`, so every
       // instance is guaranteed to implement it.
       return item[IS_DEEP_FROZEN](checkValue);
     } else if (Array.isArray(item)) {
