+++ packages/data-model/src/fabric-instances/FabricLink.ts
@@ -4,13 +4,12 @@ import type {
 } from "@commonfabric/api";
 import { isPlainObject, isUnsafeObjectKey } from "@commonfabric/utils/types";
 
+import type { FabricPlainObject, FabricValue } from "@/interface.ts";
 import {
+  BaseFabricInstance,
   DEEP_FREEZE,
-  type FabricPlainObject,
-  type FabricValue,
   IS_DEEP_FROZEN,
-} from "@/interface.ts";
-import { BaseFabricInstance } from "./BaseFabricInstance.ts";
+} from "./BaseFabricInstance.ts";
 import { cloneIfNecessary } from "@/value-clone.ts";
 import { deepFreeze, isDeepFrozen } from "@/deep-freeze.ts";
 import { BaseFabricCodec } from "@/codec-common/BaseFabricCodec.ts";
